Remove macro_ioreg crate
